requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.9
django-baton==2.9.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
exceptiongroup==1.2.0
idna==3.6
importlib-resources==6.1.1
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
model-bakery==1.17.0
packaging==23.2
pkgutil_resolve_name==1.3.10
pluggy==1.4.0
prompt-toolkit==3.0.43
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
zipp==3.17.0
